Fix circular deps, fix sigint handler
